requirements.txt
tensorflow>=2.18.0
numpy>=1.25.0
pandas>=2.1.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
fastapi>=0.111.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.32.0
streamlit>=1.30.0  # Added for the frontend application
Pillow>=10.2.0      # Added for image handling (PIL import)